Add files via upload
code to control a brushless motor, using a smartwatch. the device is detected as a common BT audio device. and by simply adjusting the volume, you adjust the speed of the motor. allows you to motorise a Sup or a kayak and control the motor from the waterproof smartwatch, without the need for additional controllers.